# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/platform/tenants/{id}": {
      "get": {
        "operationId": "TenantsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter tenant por id",
        "tags": [
          "Platform — Tenants"
        ]
      },
      "patch": {
        "operationId": "TenantsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTenantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar tenant (incl. active)",
        "tags": [
          "Platform — Tenants"
        ]
      }
    },
...
    "/platform/tenants/{tenantId}/coins": {
      "get": {
        "operationId": "CoinsController_listCoins",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar saldos de coins do tenant",
        "tags": [
          "Platform — Coins"
        ]
      }
    },
    "/platform/tenants/{tenantId}/coin-transactions": {
      "get": {
        "operationId": "CoinsController_listTransactions",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar transactions recentes do tenant",
        "tags": [
          "Platform — Coins"
        ]
      }
    },
    "/platform/tenants/{tenantId}/coins/credit": {
      "post": {
        "operationId": "CoinsController_credit",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreditCoinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Creditar coins (upsert da carteira)",
        "tags": [
          "Platform — Coins"
        ]
      }
    },
    "/platform/tenants/{tenantId}/coins/debit": {
      "post": {
        "operationId": "CoinsController_debit",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DebitCoinDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Debitar coins (amount > 0; row com amount negativo)",
        "tags": [
          "Platform — Coins"
        ]
      }
    },
    "/tenant/{tenantId}/coins": {
      "get": {
        "operationId": "TenantCoinsController_listCoins",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar saldos de coins do tenant (somente leitura)",
        "tags": [
          "Tenant — Coins"
        ]
      }
    },
    "/tenant/{tenantId}/coin-transactions": {
      "get": {
        "operationId": "TenantCoinsController_listTransactions",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar transactions recentes do tenant (somente leitura)",
        "tags": [
          "Tenant — Coins"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gestão de Leads",
    "description": "API para gestão de leads",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateTenantDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "description": "Nome do tenant"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "description": "Telefone E.164 (único)"
          },
          "active": {
            "type": "boolean",
            "description": "Tenant ativo"
          }
        }
      },
...
      "CreditCoinDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "User do tenant que receberá o crédito",
            "example": 1
          },
          "amount": {
            "type": "number",
            "minimum": 1,
            "description": "Valor positivo a creditar",
            "example": 10
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "description": "Descrição livre (marcador do operador é anexado)",
            "example": "Credito inicial onboarding"
          },
          "type": {
            "type": "string",
            "enum": [
              "CREDITO",
              "BONUS"
            ],
            "description": "Tipo da transaction (default CREDITO)",
            "default": "CREDITO"
          }
        },
        "required": [
          "userId",
          "amount"
        ]
      },
      "DebitCoinDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "User do tenant a debitar",
            "example": 1
          },
          "amount": {
            "type": "number",
            "minimum": 1,
            "description": "Valor positivo a debitar",
            "example": 3
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "description": "Descrição livre (marcador do operador é anexado)",
            "example": "Ajuste operacional"
          }
        },
        "required": [
          "userId",
          "amount"
        ]
      },
...
    }
  }
}